requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.32.4
PyMySQL==1.1.1
cryptography==44.0.1